requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.3
Django==4.0.3
djangorestframework==3.13.1
future==0.16.0
idna==3.3
jmespath==0.10.0
pathspec==0.9.0
# pypiwin32==223
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
# pywin32==303
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
wcwidth==0.1.9
